# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "서랍 API",
        "description": "Team.AM의 첫번째 프로젝트 입니다. [https://github.com/meteopark/bookmark-laravel](https://github.com/meteopark/bookmark-laravel)",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "rzip84@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/category": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "카테고리 조회",
                "description": "카테고리 조회",
                "operationId": "show",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "unauthorized token"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Category"
                ],
                "summary": "카테고리 생성",
                "description": "카테고리 생성",
                "operationId": "store",
                "requestBody": {
                    "description": "",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/BookmarkCategoryCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "unauthorized token"
                    },
                    "409": {
                        "description": "unknown user"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/category/{category_id}": {
            "post": {
                "tags": [
                    "Category"
                ],
                "summary": "카테고리명 수정",
                "description": "카테고리명 수정",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "path",
                        "description": "category_id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/BookmarkCategoryUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "unauthorized token"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Category"
                ],
                "summary": "카테고리 삭제",
                "description": "카테고리 삭제",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "path",
                        "description": "category_id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "unauthorized token"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/bookmarks": {
            "post": {
                "tags": [
                    "Bookmark"
                ],
                "summary": "북마크 생성",
                "description": "북마크 생성",
                "operationId": "store",
                "requestBody": {
                    "description": "",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/BookmarkCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "bad request"
                    },
                    "401": {
                        "description": "unauthorized token"
                    },
                    "409": {
                        "description": "unknown user"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/bookmarks/{category_id}": {
            "get": {
                "tags": [
                    "Bookmark"
                ],
                "summary": "북마크 리스트 가져오기",
                "description": "북마크 리스트 가져오기",
                "operationId": "index",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "path",
                        "description": "category_id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "unauthorized token"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/bookmarks/{category_id}/{bookmark_id}": {
            "delete": {
                "tags": [
                    "Bookmark"
                ],
                "summary": "북마크 삭제하기",
                "description": "북마크 삭제하기",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "path",
                        "description": "category_id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "bookmark_id",
                        "in": "path",
                        "description": "bookmark_id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "unauthorized token"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/bookmarks/move": {
            "post": {
                "tags": [
                    "Bookmark"
                ],
                "summary": "북마크 이동",
                "description": "북마크 이동",
                "operationId": "move",
                "requestBody": {
                    "description": "",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/BookmarkMove"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "unauthorized token"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/users": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "회원가입",
                "description": "회원가입",
                "operationId": "store",
                "requestBody": {
                    "description": "",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/UserCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "input error"
                    }
                }
            }
        },
        "/api/v1/users/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "로그인",
                "description": "로그인",
                "operationId": "login",
                "requestBody": {
                    "description": "",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/UserLogin"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "unauthorized token"
                    },
                    "409": {
                        "description": "unknown user"
                    }
                }
            }
        },
        "/api/v1/token": {
            "post": {
                "tags": [
                    "Token"
                ],
                "summary": "토큰 재발행",
                "description": "토큰 재발행",
                "operationId": "refreshToken",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "unauthorized token"
                    },
                    "409": {
                        "description": "unknown user"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "BookmarkCategoryCreate": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "BookmarkCategoryCreate"
                }
            },
            "BookmarkCategoryUpdate": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "_method": {
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "BookmarkCategoryUpdate"
                }
            },
            "BookmarkCreate": {
                "required": [
                    "category_id",
                    "url"
                ],
                "properties": {
                    "category_id": {
                        "type": "string"
                    },
                    "url": {
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "BookmarkCreate"
                }
            },
            "BookmarkMove": {
                "required": [
                    "category_id",
                    "bookmark_id"
                ],
                "properties": {
                    "category_id": {
                        "type": "string"
                    },
                    "bookmark_id": {
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "BookmarkMove"
                }
            },
            "UserCreate": {
                "required": [
                    "join_type",
                    "sns_id",
                    "profile_image",
                    "name"
                ],
                "properties": {
                    "join_type": {
                        "type": "string",
                        "enum": [
                            "kakao",
                            "facebook",
                            "google"
                        ]
                    },
                    "sns_id": {
                        "type": "string"
                    },
                    "profile_image": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "UserCreate"
                }
            },
            "UserLogin": {
                "required": [
                    "sns_id",
                    "join_type"
                ],
                "properties": {
                    "join_type": {
                        "type": "string",
                        "enum": [
                            "kakao",
                            "facebook",
                            "google"
                        ]
                    },
                    "sns_id": {
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "UserCreate"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "externalDocs": {
        "description": "Find out more about Swagger",
        "url": "http://swagger.io"
    }
}